Track mixer target output time between runs (#349)
When determining the amount of times to run the mixer, use the media end time enqueued during the previous run, instead of the wall clock time of the end of the last run. If the ticker runs late, or the mixing takes a non negligible amount of time, we would currently ignore partial frame times, potentially causing us to enqueue too little media.